Add record unit test for Language

+++ src/main/java/com/napier/sem/Record.java
@@ -124,7 +124,7 @@ public class Record
     }
 
     // Language constructor, just used for testing purposes.
-    public Record(String language, Integer speakers, Integer percentage)
+    public Record(String language, Long speakers, Integer percentage)
     {
         this.language = language;
         this.speakers = speakers;
